+++ CHANGELOG.md
@@ -7,6 +7,9 @@ The format is based on Keep a Changelog (https://keepachangelog.com), and this p
 ### Added
 - Set and Get values to/from arrays. Add relevant unit tests.
 
+### Fixed
+- Add missing unit tests for parser helper (PrintStmt).
+
 ## [Unreleased] - 2026-02-02
 ### Added
 - Add `DIM` support in Apple II Basic. Add relevant unit tests.